requirements.txt
Django==6.0
django-rq==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.12
PyJWT==2.12.1
redis==7.4.0
rq==2.8.0
Pillow==12.2.0
python-dotenv==1.2.2
